ira colocar o começo do statement

+++ src/main/java/br/com/brbank/dto/DepositDto.java
@@ -1,5 +1,5 @@
 package br.com.brbank.dto;
 
-public record DepositDto(Long money) {
+public record DepositDto(Double money) {
 
 }
