+++ src/main/java/br/com/brbank/dto/TransferDto.java
@@ -1,5 +1,5 @@
 package br.com.brbank.dto;
 
-public record TransferDto(String cpf, Long money) {
+public record TransferDto(String cpf, Double money) {
 
 }
