+++ src/main/java/br/com/brbank/dto/WithdrawDto.java
@@ -1,5 +1,5 @@
 package br.com.brbank.dto;
 
-public record WithdrawDto(Long money) {
+public record WithdrawDto(Double money) {
 
 }
